Add ChangeDonationField Component
