+++ TODO.md
@@ -1,0 +1,13 @@
+CSS: 
+- pref to tailwind
+- add charity/add creator button spacing
+- Get spinner on 1st search (Maybe some react/css) 
+- make it look a little crisper
+- Add hover animations for search results
+
+JS: 
+- Get debounce updated
+
+Potential alt tools: 
+- maybe use <DebounceInput> and <Compo> from `holaplex\src\common\components\search\SearchBar.tsx`
+- https://headlessui.dev/react/listbox
