+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@holaplex/ui",
-  "version": "0.0.992",
+  "version": "0.0.xxx",
   "description": "Holaplex UI",
   "main": "lib/index.cjs.js",
   "types": "lib/index.d.ts",
